% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemropendata.R
\docType{package}
\name{lemropendata-package}
\alias{lemropendata}
\alias{lemropendata-package}
\title{lemropendata}
\description{
lemropendata is an R interface to the LEMR Open Data Portal.
The goal of the package is to help read data directly into R without needing
to manually download it via the portal.
}
\author{
\strong{Maintainer}: Sharla Gelfand \email{sharla.gelfand@gmail.com}

}
\keyword{internal}
