% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{list_packages}
\alias{list_packages}
\title{List packages}
\usage{
list_packages(token = get_token())
}
\arguments{
\item{token}{An API key or token for the open data portal. Defaults to to the value in \code{get_token}, which can be set via \code{set_token}.}
}
\value{
A tibble of available packages and metadata.
}
\description{
List packages available on the portal.
}
\examples{
\donttest{
list_packages()
}
}
