% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{search_packages}
\alias{search_packages}
\title{Search packages by name}
\usage{
search_packages(name, token = get_token())
}
\arguments{
\item{name}{Name to search (case-insensitive)}

\item{token}{An API key or token for the open data portal. Defaults to to the value in \code{get_token}, which can be set via \code{set_token}.}
}
\value{
A tibble of matching packages along with package metadata.
}
\description{
Search portal packages by name
}
\examples{
\donttest{
search_packages("Toronto")
}
}
