% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_portal}
\alias{browse_portal}
\title{Open the LEMR Open Data Portal in your browser}
\usage{
browse_portal()
}
\description{
Open the LEMR Open Data Portal in your browser
}
\examples{
\donttest{
browse_portal()
}
}
