% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{search_packages}
\alias{search_packages}
\title{Search packages by title}
\usage{
search_packages(title, token = get_token())
}
\arguments{
\item{title}{Title to search (case-insensitive).}
}
\value{
A tibble of matching packages along with package metadata.
}
\description{
Search portal packages by title.
}
\examples{
\donttest{
search_packages("Toronto")
}
}
