% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_token}
\alias{set_token}
\alias{get_token}
\title{Get or set LEMR Open Data token}
\usage{
set_token(token, var = "LEMROPENDATA_KEY")

get_token(var = "LEMROPENDATA_KEY")
}
\arguments{
\item{token}{LEMR Open Data token}

\item{var}{Environment to store token in, defaults to \code{LEMROPENDATA_KEY}}
}
\description{
Get or set LEMR Open Data token
}
\examples{
set_token("test")
get_token()
# [1] "test"
}
